% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syncPacks.r
\name{getMyBearings}
\alias{getMyBearings}
\title{Collect basic info about your environment}
\usage{
getMyBearings(syncfolder, packsdb = NULL)
}
\arguments{
\item{syncfolder}{Character string, name of parent folder.}

\item{packsdb}{Character string, name of subfolder, if any.}
}
\description{
Collect basic info about your environment
}
\keyword{utilities}

