% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publishGithub.R
\name{publishGithub}
\alias{publishGithub}
\title{A tweak on \code{slidify::publish_github()}.}
\usage{
publishGithub(repo, username = getOption("github.user"))
}
\arguments{
\item{repo}{your public GitHub repo, e.g. "syncR"}

\item{username}{your GitHub user name, e.g. "ghuiber"}
}
\description{
Assuming that you're already working on code that's being backed up on GitHub, and that your final output is .html knitted from a .Rmd file, you might as well publish it as a \href{https://pages.github.com}{GitHub page}. This function does it with one simple call.
}

