% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/putInCalendar.R
\name{putInCalendar}
\alias{putInCalendar}
\title{Write monthly calendar for given date}
\usage{
putInCalendar(ds, offs = 0)
}
\arguments{
\item{ds}{date string of format 'YYYY-MM-DD'}

\item{offs}{offset in days: negative or positive integer number of days before or after \code{ds}}
}
\description{
Write monthly calendar for given date
}

