% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syncPacks.r
\name{syncPacks}
\alias{syncPacks}
\title{Syncs packages across your computers}
\usage{
syncPacks(syncfolder = "SpiderOak Hive", packsdb = "syncR")
}
\arguments{
\item{syncfolder}{Character string, name of parent folder.}

\item{packsdb}{Character string, name of subfolder, if any.}
}
\description{
Prerequisites: you must keep your R packages listed
in a synced folder, such as Dropbox or SpiderOak Hive.
The packages themselves, that is, are in you \code{lib}
folder. What you keep in the synced folder is only
an \code{.RData} file that lists them by name.
This is your package database, one per computer. To
keep these databases separated from whatever else
you have in the synced folder, give them a subfolder
of their own, referenced here as \code{packsdb}.
}
\examples{
\dontrun{
syncPacks('SpiderOak Hive','syncR')
syncPacks()
}
}
\keyword{utilities}

