% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syncR-package.r
\docType{package}
\name{syncR}
\alias{syncR}
\alias{syncR-package}
\title{syncR.}
\description{
Sync R packages across your computers.
To see what's here, type \code{ls('package:syncR')}
at the console.
}

