% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncr.R
\name{syncr}
\alias{syncr}
\title{syncr}
\usage{
syncr(src, dest, archive = TRUE, compress = TRUE, verbose = FALSE,
  relative = FALSE, delete = FALSE, dry_run = FALSE,
  drop_src_directory = FALSE, args_only = FALSE)
}
\arguments{
\item{src}{Source files to copy.  See details}

\item{dest}{A single destination path, possibly remote.}

\item{archive}{Copy in archive mode, being recursive, preserving
attributes etc?  Generally this is wanted.}

\item{compress}{Compress files on transfer?}

\item{verbose}{Be verbose? (Always \code{TRUE} when using \code{dry_run}).}

\item{relative}{Copy relative paths only, rather than the last
part of the file name.  So if you have a directory \code{src}
and you want to syncronise all the \code{.c} files in it, you
could use \code{src="src/*.c", relative=TRUE}.  You can include
a dot at any point in a path name to indicate where the relative
directory should start from;
e.g. \code{/absolute/path/to/./src/*.c} which will create a
directory \code{src} with .c files in it.}

\item{delete}{Delete files on the remote machine.  When mirroring
this is generally what you want to do but it is not enabled by
default because it's potentially dangerous.}

\item{dry_run}{Don't actually do anything, but print what would be
done instead.}

\item{drop_src_directory}{In the case where \code{src} is a single
directory, don't copy the directory, but copy the contents.}

\item{args_only}{Don't run anything and instead return the
arguments that would have been passed to rsync.}
}
\description{
Wrapper around rsync, with only a few options exposed.
\emph{WARNING}: This command modifies files on your computer and
therefore is potentially dangerous.  Files in \code{dest} may be
overwritten or deleted!
}
\details{
This is a simple wrapper around a few of the most common arguments
to \code{rsync}.  For a more complete interface, see
\code{\link{rsync}}.
}

